Everything as expected; no damping
When there is no damping, the system will begin to diverge when the restoring force is decaying over time. Thus, the rocket's path is stable while it is under thrust, but it becomes unstable once it stops accelerating. When it re-enters the low atmosphere on descent, it becomes stable once more. Hopefully aerodynamic damping will provide stability through the flight.